move stats renderer from main module to angular example
