+++ example/before-evening/src/app/car-game/car-game.component.ts
@@ -1,7 +1,8 @@
-import {AfterViewInit, Component} from '@angular/core';
-import {BeforeEvening} from "../../../../../build/main";
-import {GameStateService} from "../game-state.service";
-import {StateUpdate} from "../../../../../build/main/interfaces/state.interfaces";
+import { AfterViewInit, Component } from '@angular/core';
+import { BeforeEvening } from '../../../../../build/main';
+import { GameStateService } from '../game-state.service';
+import { StateUpdate } from '../../../../../build/main/interfaces/state.interfaces';
+import { StatsRenderer } from './statsRenderer';
 
 @Component({
   selector: 'car-game',
@@ -20,8 +21,13 @@ export class CarGameComponent implements AfterViewInit {
     this.gameStateService.associateStateUpdater(beforeEvening.stateUpdate);
     beforeEvening.runGame();
 
+    const statsContainer = new StatsRenderer(beforeEvening.stats);
+
     beforeEvening.stateUpdate.subscribe(state => {
       this.currentState = state;
+      statsContainer.updateStats(beforeEvening.stats.ms, beforeEvening.stats.msMin,
+        beforeEvening.stats.msMax, beforeEvening.stats.fps,
+        beforeEvening.stats.fpsMin, beforeEvening.stats.fpsMax);
     });
 
     this.gameStateService.refreshGame.subscribe(() => {
